Coefficient of regression

coefficient_of_regression calculate variance of a genotype across environments.

coefficient_of_regression(data, trait, genotype, environment)

Arguments

data

a dataframe containing trait, genotype and environment.

trait

colname of a column containing a numeric vector of interested trait to be analysized.

genotype

colname of a column containing a character or factor vector labeling different genotypic varieties

environment

colname of a column containing a character or factor vector labeling different environments

Value

a data table with coefficient of regression

Details

Coefficient of regression (Finlay and Wilkinson, 1963) is calculatd based on regression function. Variety with low coefficient of regression is considered as stable. Under the linear model $$Y =\mu + b_{i}e_{j} + g_{i} + d_{ij}$$ where Y is the predicted phenotypic values, \(g_{i}\), \(e_{j}\) and \(\mu\) denoting genotypic, environmental and overall population mean,respectively.

The effect of GE-interaction may be expressed as: $$(ge)_{ij} = b_{i}e_{j} + d_{ij}$$ where \(b_{i}\) is the coefficient of regression and \(d_{ij}\) a deviation.

Coefficient of regression may be expressed as: $$ b_{i}=1 + \frac{\sum_{j} (X_{ij} -\bar{X_{i.}}-\bar{X_{.j}}+\bar{X_{..}})\cdot (\bar{X_{.j}}- \bar{X_{..}})}{\sum_{j}(\bar{X_{.j}}-\bar{X_{..}})^{2}}$$

where \(X_{ij}\) is the observed phenotypic mean value of genotype i(i=1,..., G) in environment j(j=1,...,E), with \(\bar{X_{i.}}\) and \(\bar{X_{.j}}\)
denoting marginal means of genotype i and environment j,respectively.
\(\bar{X_{..}}\) denote the overall mean of X.

References

Finlay KW, Wilkinson GN (1963). “The analysis of adaptation in a plant-breeding programme.” Australian Journal of Agricultural Research, 14(6), 742--754. doi: 10.1071/AR9630742 .

Author

Tien Cheng Wang

Examples

data(Data)
coefficient.of.regression <- coefficient_of_regression(
 data = Data,
 trait = "Yield",
 genotype = "Genotype",
 environment = "Environment")