Table of stability indices

table_stability export all the stability indices in the package.

table_stability(
  data,
  trait,
  genotype,
  environment,
  lambda,
  normalize = FALSE,
  unit.correct = FALSE
)

Arguments

data
a data frame containing trait, genotype and environment.

trait
colname of a column containing a numeric vector of interested trait to be analyzed.

genotype
colname of a column containing a character or factor vector labeling different genotypic varieties.

environment
colname(s) of a column containing a character or factor vector labeling different environments, if input is a vector containing multiple column names, then it will be merged into single environment column in the function.

lambda
the minimal acceptable value of trait that the user expected from crop across environments. Lambda should between the ranges of trait vlaue.

normalize
logical, default is FALSE, indicating whether stability indices should be normalized to the range from 0 to 1, where 1 refer to stable and 0 is unstable.

unit.correct
logical, default is FALSE, returning the stability index with unit equals to squared unit of trait; when TRUE, returning stability index with the unit as same as unit of trait.

Value

a data table with multiple stability indices

Details

Combine all stability indices in this package and export as a table, including mean trait, normality of the trait across environment.

References

Doering TF, Reckling M (2018). “Detecting global trends of cereal yield stability by adjusting the coefficient of variation.” European Journal of Agronomy, 99, 30 - 36. doi: 10.1016/j.eja.2018.06.007 , http://www.sciencedirect.com/science/article/pii/S1161030118301904. Pinthus MJ (1973). “Estimate of genotypic value: A proposed method.” Euphytica, 22, 121--123. doi: 10.1007/BF00021563 . Finlay KW, Wilkinson GN (1963). “The analysis of adaptation in a plant breeding programme.” Australian Journal of Agricultural Research, 14, 742--754. Eberhart SA, Russell WA (1966). “Stability parameters for comparing varieties 1.” Crop science, 6, 36--40. Wricke G (1962). “Ueber eine Methode zur Erfassung der 昼㸶kologischen Streubreite in Feldversuchen.” Z. Pflanzenzuechtg, 47, 92--96. Roemer T (1917). “Sind die ertragreichen Sorten ertragsicherer?” Mitt DLG, 32, 87--89. Hanson WD (1970). “Genotypic stability.” Theoretical and Applied Genetics, 40, 226--231. doi: 10.1007/BF00285245 . Lin CS, Binns MR (1988). “A superiority measure of cultivar performance for cultivar 搼㸷 location data.” Canadian Journal of Plant Science, 68, 193-198. doi: 10.4141/cjps88-018 . Shukla GK (1972). “Some statistical aspects of partitioning genotype environmental components of variability.” Heredity, 29, 237--245. Nassar R, Huehn M (1987). “Studies on estimation of phenotypic stability: Tests of significance for nonparametric measures of phenotypic stability.” Biometrics, 45--53. Eskridge KM (1990). “"Selection of Stable Cultivars Using a Safety-First Rule".” Crop Science, 30, 369 - 374. doi: 10.2135/cropsci1990.0011183X003000020025x .

See also

Author

Tien-Cheng Wang

Examples

data(Data)
tb <- table_stability(
 data = Data,
 trait = "Yield",
 genotype = "Genotype",
 environment = "Environment",
 lambda = median(Data$Yield),
 normalize = TRUE,
 unit.correct=TRUE)

#> Warning: All of your genotypes didn't pass the Shapiro normality test!
#>  Safety_first Index may not be accurate.